APP/requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
plotly>=5.24.0
scipy>=1.14.0
numpy>=2.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
